requirements.txt
blinker==1.9.0
click==8.3.1
dlib==20.0.0
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.0
imutils==0.5.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==1.24.3
opencv-python==4.8.1.78
openpyxl==3.1.2
pandas==2.1.4
Pillow==10.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.10.1
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.4
